requirement.txt
fastapi[standard]==0.121.0
pyjwt==2.10.1
sqlmodel==0.0.27
sqlalchemy==2.0.44
python_dotenv==1.2.1
pwdlib[argon2]==0.3.0
